Move helpers to separate file
